refactor: 컴포넌트 내부로 useContext hook 이동
